Bluetooth: Add Thermometer profile and integrate changes to fix few of the reported issues
List of the fixes included:
--
commit 1a523a174ee1429a46462a82375d4c75012985f8
Author: Archana Ramachandran<archanar@codeaurora.org>

Bluetooth: Thermometer Profile Implementation
--
commit 2d9e4ded927296b92b2fa5e7c97efdf91ddf66f0
Author: Archana Ramachandran<archanar@codeaurora.org>

Bluetooth: Thermometer client profile service
--
commit fffafb6c14fb1298cb2955168a0af893ae7a4fcf
Author: Ankit Verma<ankitv@codeaurora.org>

Bluetooth: Fixed klocwork issues
--
commit 22648a418ebd04216a183364c5d5fe4168a5745c
Author: Priyanka Vunnam<pvunnam@codeaurora.org>

Bluetooth: Fixed klocwork issues.
--
commit fb683bfa6903aff4f66dea684b168dd831070357
Author: Archana Ramachandran<archanar@codeaurora.org>

Fix klocwork issues
--
commit 54de9e87f4befae1d3dc89aebf00381c161bf1a7
Author: Archana Ramachandran<archanar@codeaurora.org>

Adding Device information gatt service to the thermometer profile
--
commit 4d74bc3f9e483235fd8403899c3180ce7ef79b53
Author: Vadivel Thekkamalai<vthekk@codeaurora.org>

Bluetooth: Fix and code optimization for FTP folder listing

CRs-fixed: 310637
--
commit 5dd1f121bd734c78f74df7763cff98ec39560927
Author: Kannu Mehta<kmehta@codeaurora.org>

Bluetooth: Case insensitive sorting in PBAP.

CRs-fixed: 315654
--
commit d12f3c5f0018e6508144050d20ce0ac520acd8cc
Author: Kun Han Kim<kunhank@codeaurora.org>

Bluetooth: Force close issue

CRs-fixed: 315855
--
commit 1a5f39e18e5aafbd57d256a3c06431d81d469f8b
Author: Inga Stotland<ingas@codeaurora.org>

Bluetooth: Changes for new Gatt WriteCharacteristicRaw method.
--
commit 1975ccdbfd3cad716cb25d0ee044ee4848edb8ab
Author: Vadivel Thekkamalai<vthekk@codeaurora.org>

Bluetooth: Reject second FTP connection with the PSM/SCN which
is already in use.

CRs-fixed: 317042
--
commit e2600c873c52ab02d5dbf4d938e5d5a02049c3d5
Author: Kannu Mehta<kmehta@codeaurora.org>

Bluetooh: PBAP: Get contacts by numbers.

CRs-fixed: 316506
--
commit ef5fa47c821ecf0fcb94a5a1dc165a7ccfcc96c7
Author: Vadivel Thekkamalai<vthekk@codeaurora.org>

Bluetooth: Restrict FTP operations on directory name . and ..

CRs-fixed: 321589
--

Conflicts:

	src/com/android/bluetooth/opp/BluetoothOppService.java

+++ src/com/android/bluetooth/ftp/FileUtils.java
@@ -37,6 +37,7 @@ import android.os.Bundle;
 import android.webkit.MimeTypeMap;
 
 import java.io.IOException;
+import java.io.FileNotFoundException;
 import java.io.OutputStream;
 import java.io.InputStream;
 import java.io.File;
@@ -71,6 +72,10 @@ public class FileUtils {
         if (D) Log.d(TAG, "deleteDirectory() +");
         if(dir.exists()) {
             File [] files = dir.listFiles();
+            if (files == null) {
+                Log.e(TAG, "error in listing directory ");
+                return false;
+            }
             for(int i = 0; i < files.length;i++) {
                 if(files[i].isDirectory()) {
                     deleteDirectory(mCallback,files[i]);
@@ -105,6 +110,10 @@ public class FileUtils {
          int ret = 0;
          dest.mkdir();
          File [] files = src.listFiles();
+         if (files == null) {
+             Log.e(TAG, "error in listing directory");
+             return ResponseCodes.OBEX_HTTP_INTERNAL_ERROR;
+         }
          for(int i = 0; i < files.length; i++) {
              if (D) Log.d(TAG,"Files =" + files[i]);
              if(files[i].isDirectory()) {
@@ -143,9 +152,21 @@ public class FileUtils {
         try {
             reader = new FileInputStream(src);
             writer = new FileOutputStream(dest);
+        } catch(FileNotFoundException e) {
+            Log.e(TAG,"copyFile file not found "+ e.toString());
+            return ResponseCodes.OBEX_HTTP_INTERNAL_ERROR;
         } catch(IOException e) {
             Log.e(TAG,"copyFile open stream failed "+ e.toString());
             return ResponseCodes.OBEX_HTTP_INTERNAL_ERROR;
+        } finally {
+            if (null != reader && null == writer) {
+                try {
+                    reader.close();
+                } catch (IOException e) {
+                    Log.e(TAG, "copyFile close stream failed"+ e.toString());
+                    return ResponseCodes.OBEX_HTTP_INTERNAL_ERROR;
+                }
+            }
         }
 
         BufferedInputStream ins = new BufferedInputStream(reader, 0x40000);
